Relative Nodes simulation
