+++ modules/SpreadSimulator/src/main/java/Components/Simulation/SimulationButton.java
@@ -1,5 +1,7 @@
 package Components.Simulation;
 
+import Components.Simulation.Simulation.Simulation;
+import Components.Simulation.Simulation.SimulationAll;
 import ConfigLoader.ConfigLoader;
 import javax.swing.*;
 import java.awt.*;
