+++ modules/SpreadSimulator/src/main/java/Components/Simulation/SimulationComponent.java
@@ -1,5 +1,8 @@
 package Components.Simulation;
 
+import Components.Simulation.Simulation.Simulation;
+import Components.Simulation.Simulation.SimulationAll;
+import Components.Simulation.Simulation.SimulationRelativeNodes;
 import ConfigLoader.ConfigLoader;
 import Helper.ApplySimulationHelper;
 import Helper.ObjectMapperHelper;
@@ -7,7 +10,6 @@ import SimulationModel.Interaction.RelativeEdgesInteraction;
 import SimulationModel.Interaction.RelativeFreeEdgesInteraction;
 import SimulationModel.Interaction.RelativeFreeNodesInteraction;
 import SimulationModel.Interaction.RelativeNodesInteraction;
-import SimulationModel.Node.NodeRole;
 import SimulationModel.Node.NodeRoleDecorator;
 import SimulationModel.Node.NodeStateDecorator;
 import SimulationModel.SimulationModel;
@@ -80,7 +82,22 @@ public class SimulationComponent extends TopComponent {
                 var path = new File(ConfigLoader.folderSimulationTmp + ConfigLoader.folderSimulationTmpFilename);
                 var content = new String(Files.readAllBytes(Paths.get(path.getAbsolutePath())));
                 var simulationModel = mapper.readValue(content, SimulationModel.class);
-                simulation = new Simulation(graph, simulationModel);
+                switch (simulationModel.getInteraction().getInteractionType()){
+                    case All:
+                        simulation = new SimulationAll(graph, simulationModel);
+                        break;
+                    case RelativeEdges:
+                        break;
+                    case RelativeFreeEdges:
+                        break;
+                    case RelativeNodes:
+                        simulation = new SimulationRelativeNodes(graph, simulationModel);
+                        break;
+                    case RelativeFreeNodes:
+                        break;
+                    default:
+                        break;
+                }
                 initComponents();
                 revalidate();
                 repaint();
@@ -128,7 +145,7 @@ public class SimulationComponent extends TopComponent {
 
         int row = 1;
         int padding = 4;
-        for (NodeRoleDecorator role : simulation.nodeRoleDecoratorList) {
+        for (NodeRoleDecorator role : simulation.getNodeRoleDecoratorList()) {
             addRoleToPanel(panel, role, gbc, row, padding);
             row += 3;
             for (NodeStateDecorator state : role.getNodeStates()) {
